# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreateSaleCenter:
    properties:
      branch:
        type: string
      name:
        type: string
    type: object
...
  models.SaleCenter:
    properties:
      branch:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.SaleCenterGetListResponse:
    properties:
      count:
        type: string
      sale_centers:
        items:
          $ref: '#/definitions/models.SaleCenter'
        type: array
    type: object
...
  models.UpdateSaleCenter:
    properties:
      branch:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://demo.com/support
  description: Create  Go REST API with JWT Authentication in Gin Framework
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: demo.com
  title: Swagger JWT API
  version: "1.0"
paths:
...
  /center:
    get:
      consumes:
      - application/json
      description: Get List Center
      operationId: get_list_center
      parameters:
      - description: limit
        in: query
        name: limit
        type: number
      - description: offset
        in: query
        name: offset
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: GetListCenterResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SaleCenterGetListResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Center
      tags:
      - Center
    post:
      consumes:
      - application/json
      description: Create Center
      operationId: create_center
      parameters:
      - description: CreateCenterRequestBody
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.CreateSaleCenter'
      produces:
      - application/json
      responses:
        "200":
          description: CenterBody
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SaleCenter'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Center
      tags:
      - Center
  /center/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Center
      operationId: delete_center
      parameters:
      - description: DeleteCenterPath
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted Center
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Center
      tags:
      - Center
    get:
      consumes:
      - application/json
      description: Get By Id Center
      operationId: get_by_id_center
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: CenterBody
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SaleCenter'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By Id Center
      tags:
      - Center
    put:
      consumes:
      - application/json
      description: Update Center
      operationId: update_center
      parameters:
      - description: CenterPrimaryKey_ID
        in: path
        name: id
        required: true
        type: string
      - description: UpdateCenterBody
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSaleCenter'
      produces:
      - application/json
      responses:
        "200":
          description: Updated Center
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Center
      tags:
      - Center
...
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"